rename: 가독성을 위한 bestProductList -> bestProducts 이름 변경
